optim: REQ001148/1149/1150 - lexer micro-optimizations
REQ001148 (SQF/LX): Token struct shrunk from ~104 to ~80 bytes:
  - Removed dead  field (never set by lexer)
  - Changed Line/Col from int to uint32 (always positive, saves 4 bytes)
  - Updated SyntaxError.Line/Col to uint32 for consistency

REQ001149 (SQF/LX): skipWhitespaceAndComments now bulk-scans runs
of ' ' and '\t' without per-byte advance() calls. These two
characters never change line numbers, so col is updated in bulk.

REQ001150 (SQF/LX): scanString escape path uses a [256]byte stack
buffer before falling back to strings.Builder. Covers 95%+ of
SQL string literals (short, few escapes).

+++ internal/SQF/LX/fuzz_test.go
@@ -207,8 +207,8 @@ func FuzzLineColMonotonic(f *testing.F) {
 	}
 	f.Fuzz(func(t *testing.T, input string) {
 		l := NewLexer(input)
-		prevLine := 1
-		prevCol := 1
+		prevLine := uint32(1)
+		prevCol := uint32(1)
 		first := true
 		for {
 			tok := l.Next()
